pyproject.toml
[project]
name = "legal-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
